MeshComponent and MaterialComponent are added to Entity separately

+++ Source/Graphics/Color.cpp
@@ -4,6 +4,6 @@ Color::Color() : color(glm::vec4(1.0, 1.0, 1.0, 1.0)) {
 
 }
 
-Color::Color(int r, int g, int b, int a) : color(glm::vec4(r, g, b, a)) {
+Color::Color(float r, float g, float b, float a) : color(glm::vec4(r, g, b, a)) {
 
 }
